feat(nest, contract): prefer import populateContractRouterPaths from @orpc/contract
Fixes #1311
