+++ apps/content/docs/openapi/integrations/implement-contract-in-nest.md
@@ -64,8 +64,7 @@ Before implementation, define your oRPC contract. This process is consistent wit
 ::: details Example Contract
 
 ```ts
-import { populateContractRouterPaths } from '@orpc/nest'
-import { oc } from '@orpc/contract'
+import { oc, populateContractRouterPaths } from '@orpc/contract'
 import * as z from 'zod'
 
 export const PlanetSchema = z.object({
