+++ packages/nest/src/utils.test.ts
@@ -1,6 +1,4 @@
-import { oc } from '@orpc/contract'
-import { inputSchema, outputSchema } from '../../contract/tests/shared'
-import { populateContractRouterPaths, toNestPattern } from './utils'
+import { toNestPattern } from './utils'
 
 it('toNestPattern', () => {
   expect(toNestPattern('/ping')).toBe('/ping')
@@ -9,30 +7,4 @@ it('toNestPattern', () => {
   expect(toNestPattern('/{id}/{+path}')).toBe('/:id/*path')
 
   expect(toNestPattern('/{id}/name{name}')).toBe('/:id/name{name}')
-})
-
-it('populateContractRouterPaths', () => {
-  const contract = {
-    ping: oc.input(inputSchema),
-    pong: oc.route({
-      path: '/pong/{id}',
-    }),
-    nested: {
-      ping: oc.output(outputSchema),
-      pong: oc.route({
-        path: '/pong2/{id}',
-      }),
-    },
-  }
-
-  const populated = populateContractRouterPaths(contract)
-
-  expect(populated.pong['~orpc'].route.path).toBe('/pong/{id}')
-  expect(populated.nested.pong['~orpc'].route.path).toBe('/pong2/{id}')
-
-  expect(populated.ping['~orpc'].route.path).toBe('/ping')
-  expect(populated.ping['~orpc'].inputSchema).toBe(inputSchema)
-
-  expect(populated.nested.ping['~orpc'].route.path).toBe('/nested/ping')
-  expect(populated.nested.ping['~orpc'].outputSchema).toBe(outputSchema)
 })
